Split out scanner service into reader and writer
This was a service that really should have been two services, and the
need for the goroutine in the service should've been a hint. A panic
thrown inside a goroutine will not propagate outwards to the panic
handler, so it would kill the whole process. No services should contain
goroutines anyway as that's a sign that they're doing too much.