*Vastly* speed up integration tests
In the before times, each test package had to run in serial, because
they all shared a test db. Now, each test package gets its own db
with a random name, and drops it at the end, so they can run in parallel.

Also, each test function within a test package needed a fresh db as it
made changes and invalidated it for the next test. Now, each
test function runs in a new transaction and rolls it back at the end,
preserving the pristine db state for each subsequent test.